+++ product_code/matriculation_service/CHANGELOG.md
@@ -1,3 +1,9 @@
+# [v0.13.1](https://github.com/upb-uc4/University-Credits-4.0/compare/matriculation-v0.12.1...matriculation-v0.13.1) (2020-11-25)
+## Feature
+## Refactor
+ - Rebuilt to be compatible with new user-service API
+## Bugfix
+
 # [v0.12.1](https://github.com/upb-uc4/University-Credits-4.0/compare/matriculation-v0.12.0...matriculation-v0.12.1) (2020-11-17)
 ## Feature
  - Added configurable timeouts in application config